+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CRUDlex Changelog
 Released: Upcoming
 - Fixed a crash where the namefield of a many relation is a MySQL keyword, thanks to https://github.com/th-lange
 - Fixed a timeout within the fetching of the language names
+- Fixed a crash if a referenced entity had hard deletion
 
 ## 0.14.0
 Released: 2018-09-07